Moving setNewFolder method to modals.
