+++ src/js/messaging/actions/modals.js
@@ -1,6 +1,7 @@
 export const TOGGLE_CONFIRM_DELETE = 'TOGGLE_CONFIRM_DELETE';
 export const TOGGLE_ATTACHMENTS = 'TOGGLE_ATTACHMENTS';
 export const TOGGLE_CREATE_FOLDER = 'TOGGLE_CREATE_FOLDER';
+export const SET_FOLDER_NAME = 'SET_FOLDER_NAME';
 
 
 export function toggleConfirmDelete() {
@@ -19,5 +20,13 @@ export function toggleAttachmentsModal() {
 export function toggleCreateFolderModal() {
   return {
     type: TOGGLE_CREATE_FOLDER
+  };
+}
+
+
+export function setNewFolderName(folderName) {
+  return {
+    type: SET_FOLDER_NAME,
+    folderName
   };
 }
