+++ src/js/messaging/reducers/modals.js
@@ -1,9 +1,11 @@
 import set from 'lodash/fp/set';
 
+import { makeField } from '../../common/model/fields.js';
 import {
   TOGGLE_CONFIRM_DELETE,
   TOGGLE_ATTACHMENTS,
-  TOGGLE_CREATE_FOLDER
+  TOGGLE_CREATE_FOLDER,
+  SET_FOLDER_NAME
 } from '../actions/modals';
 
 const initialState = {
@@ -14,7 +16,8 @@ const initialState = {
     visible: false
   },
   createFolder: {
-    visible: false
+    visible: false,
+    newFolderName: makeField('')
   }
 };
 
@@ -26,6 +29,8 @@ export default function modals(state = initialState, action) {
       return set('attachments.visible', !state.attachments.visible, state);
     case TOGGLE_CREATE_FOLDER:
       return set('createFolder.visible', !state.createFolder.visible, state);
+    case SET_FOLDER_NAME:
+      return set('createFolder.folderName', action.folderName, state);
     default:
       return state;
   }
